Add a backend for timestats
This one is so small, I feel silly to have made a package just for
it. The `package.yaml` is bigger than the Haskell code! Timestats
already has support for pure code, but doesn't support evaluation
strategies. So this is just _one_ function to measure the time for a
call to `eval`.