Completely redesign the API around explicit strategies
The API is much nicer. Type families in the `Eval` type class were
getting in the way of derivation (derivation isn't done yet). It just
created a lot of boiler place. Both at call sites where we had to wrap
random values in a newtype all the time, and when defining `Eval`
instances (which may be used just once, which was entirely
unnecessary), since the instances were almost entirely boilerplate.

This actually deletes lines of code, so even the implementation is
simpler really. I don't need some special support for the
`strict-wrapper` library: just use the `seq` strategy.

This is looking quite good.

Closes #7.
Closes #17.